requirements.txt
fastapi==0.112.0
uvicorn==0.30.3
firebase-admin==6.5.0
pydantic==2.8.2
google-cloud-firestore==2.16.1
